src/telran/measure/Length.java
package telran.measure;

public class Length {
	
	private float amount;
	private LengthUnit lengthUnit;
	
	public Length(float amount, LengthUnit lengthUnit) {
		this.amount = amount;
		this.lengthUnit = lengthUnit;
	}
	
	@Override
	/**
	 * equals two Length objects according to LengthUnit and amount
	 * 10m == 10000mm
	 */
	public boolean equals(Object obj) {
		return compareTo(this) - compareTo((Length)obj) == 0;
	}
	
	/**
	 * 
	 * @param o
	 * @return < 0 "this" object less than "o" object,
	 *  > 0 "this" object greater than "o" object,
	 *  == 0 "this" object equals "o" object,
	 */
	public int compareTo(Length o) {
		return (int) (getAmount() * lengthUnit.getValue() - o.getAmount() * o.lengthUnit.getValue());
	}
	
	/**
	 * 
	 * @param unit
	 * @return new Length object with a given LengthUnit
	 * example, convert(LengthUnit.M) returns Length in meters 
	 */
	public Length convert(LengthUnit unit) {
		amount = getAmount() * getUnit().getValue() / unit.value;
		lengthUnit = unit;
		return this;
	}
	
	@Override
	/**
	 * returns string with amount and length unit pinned to amount with no space
	 * Example: 20.0M (string expression of Length object presenting 20 meters)
	 */
	public String toString() {
		return String.valueOf(getAmount()) + getUnit();
	}

	public float getAmount() {
		return amount;
	}

	public LengthUnit getUnit() {
		return lengthUnit;
	}
}
